// src/shared/enums/gameActionsEnum.ts
export enum GameAction {
    ActionStart = "Start game",
    ActionStop = "Action Stop",
    ActionUp = "Snake Go Up",
    ActionDown = "Snake Go Down",
    ActionRight = "Snake Go Right",
    ActionLeft = "Snake Go Left",
    ActionGrow = "Snake growing",
    ActionGameOver = "Game end with score: ",
    ActionResetGame = "Reset Game"
}